Split out phase1 of speedy compilation in preparation for making it stack-safe.
changelog_begin
changelog_end
